docs: Added the free PSX Shader Asset
### Commit Description:
This commit adds the shader asset and its corresponding files to the repository. This will enable our game to have the PSX look and feel throughout the entirety of the gameplay. This commit additionally adds the TMP Text Add-on for text-related Game Objects to be added onto the game scenes.

### Details:
- Added the free PSX Shader Asset available on [the Unity Store](https://assetstore.unity.com/packages/vfx/shaders/psx-shader-kit-183591)
- Sent a `README.rtf` file to the *scratch-notes* channel on the development Discord text channel.
- Added the [TextMeshPro](https://docs.unity3d.com/Manual/com.unity.textmeshpro.html) Assets to the repository, enabling UI-related functionalities to be added in future development.

### Related Issue(s):
- Closes #3